// src/types/dashboard.ts
export interface DashboardStat {
  title: string;
  value: string;
  trend: string;
  icon: any;
}

export interface DashboardEvent {
  title: string;
  date: string;
  location: string;
  image: string;
  cta: string;
}

export interface DashboardAnnouncement {
  title: string;
  icon: string;
  color: string;
  date: string;
  desc: string;
}